Basic State code lab
This codelab explains the core concepts related to using State in Jetpack Compose. It shows you how the app's state determines what is displayed in the UI, how Compose updates the UI when state changes by working with different APIs, how to optimize the structure of our composable functions, and using ViewModels in a Compose world.

https://developer.android.com/codelabs/jetpack-compose-state?continue=https%3A%2F%2Fdeveloper.android.com%2Fcourses%2Fpathways%2Fjetpack-compose-for-android-developers-1%3F_gl%3D1*19th1im*_up*MQ..%26gclid%3DCjwKCAiA2cu9BhBhEiwAft6IxGmpDZgXdoczZZNWKPZpnLiWzQDr30dGQvluGic59q6OY0mxSIaEwxoCI3oQAvD_BwE%26gclsrc%3Daw.ds%23codelab-https%3A%2F%2Fdeveloper.android.com%2Fcodelabs%2Fjetpack-compose-state&_gl=1*1n7m7ub*_up*MQ..&gclid=CjwKCAiA2cu9BhBhEiwAft6IxGmpDZgXdoczZZNWKPZpnLiWzQDr30dGQvluGic59q6OY0mxSIaEwxoCI3oQAvD_BwE&gclsrc=aw.ds#0